removed DriveSubsystem.kt and add the followTrajectory method in the SwerveDrive.kt
