+++ src/main/kotlin/frc/team449/subsystems/drive/differential/DifferentialDrive.kt
@@ -10,7 +10,6 @@ import edu.wpi.first.math.kinematics.DifferentialDriveKinematics
 import edu.wpi.first.math.kinematics.DifferentialDriveWheelSpeeds
 import edu.wpi.first.wpilibj2.command.SubsystemBase
 import frc.team449.subsystems.RobotConstants
-import frc.team449.subsystems.drive.DriveSubsystem
 import frc.team449.system.AHRS
 import frc.team449.system.encoder.Encoder
 import frc.team449.system.encoder.QuadEncoder
@@ -34,7 +33,7 @@ open class DifferentialDrive(
   private val feedForward: DifferentialDriveFeedforward,
   private val makeSidePID: () -> PIDController,
   private val trackwidth: Double
-) : DriveSubsystem, SubsystemBase() {
+) : SubsystemBase() {
   init {
     leftEncoder.resetPosition(0.0)
     rightEncoder.resetPosition(0.0)
@@ -69,7 +68,7 @@ open class DifferentialDrive(
   private var rightVel = 0.0
 
   /** Calculate left and right side speeds from given [ChassisSpeeds]. */
-  override fun set(desiredSpeeds: ChassisSpeeds) {
+  fun set(desiredSpeeds: ChassisSpeeds) {
     prevWheelSpeeds = desiredWheelSpeeds
 
     prevLeftVel = desiredWheelSpeeds.leftMetersPerSecond
@@ -99,17 +98,13 @@ open class DifferentialDrive(
   }
 
   /** The (x, y, theta) position of the robot on the field. */
-  override var pose: Pose2d
+  var pose: Pose2d
     get() = this.poseEstimator.estimatedPosition
     set(pose) {
       this.poseEstimator.resetPosition(ahrs.heading, leftEncoder.position, rightEncoder.position, pose)
     }
 
-  override fun driveFieldRelative(speeds: ChassisSpeeds) {
-    TODO("Not yet implemented")
-  }
-
-  override fun stop() {
+  fun stop() {
     this.set(ChassisSpeeds(0.0, 0.0, 0.0))
   }
