+++ src/main/kotlin/frc/team449/subsystems/drive/swerve/SwerveDrive.kt
@@ -1,5 +1,10 @@
 package frc.team449.subsystems.drive.swerve
 
+import choreo.trajectory.SwerveSample
+import edu.wpi.first.math.MathUtil
+import edu.wpi.first.math.controller.PIDController
+import edu.wpi.first.math.geometry.Pose2d
+import edu.wpi.first.math.geometry.Rotation2d
 import edu.wpi.first.math.geometry.Translation2d
 import edu.wpi.first.math.kinematics.ChassisSpeeds
 import edu.wpi.first.math.kinematics.SwerveDriveKinematics
@@ -29,6 +34,22 @@ open class SwerveDrive(
   protected val field: Field2d,
   val maxModuleSpeed: Double
 ) : SubsystemBase() {
+  private val xController: PIDController
+    get() = PIDController(10.0, 0.0, 0.0)
+  private val yController: PIDController
+    get() = PIDController(10.0, 0.0, 0.0)
+  private val headingController: PIDController
+    get() = PIDController(7.5, 0.0, 0.0)
+
+  var heading: Rotation2d
+    get() = Rotation2d(MathUtil.angleModulus(this.pose.rotation.radians))
+    set(value) {
+      this.pose = Pose2d(Translation2d(this.pose.x, this.pose.y), value)
+    }
+
+  var pose: Pose2d = Pose2d()
+
+  private var desiredSpeeds: ChassisSpeeds = ChassisSpeeds()
 
   /** The kinematics that convert [ChassisSpeeds] into multiple [SwerveModuleState] objects. */
   val kinematics = SwerveDriveKinematics(
@@ -38,11 +59,9 @@ open class SwerveDrive(
   /** The current speed of the robot's drive. */
   var currentSpeeds = ChassisSpeeds()
 
-  var desiredSpeeds: ChassisSpeeds = ChassisSpeeds()
-
   protected var speedMagnitude: Double = 0.0
 
-  fun set(desiredSpeeds: ChassisSpeeds) {
+  fun driveFieldRelative(desiredSpeeds: ChassisSpeeds) {
     this.desiredSpeeds = desiredSpeeds
 
     // Converts the desired [ChassisSpeeds] into an array of [SwerveModuleState].
@@ -64,16 +83,29 @@ open class SwerveDrive(
       module.update()
   }
 
+  fun followTrajectory(sample: SwerveSample) {
+    // Get the current pose of the robot
+
+    // Generate the next speeds for the robot
+    val speeds: ChassisSpeeds = ChassisSpeeds(
+      sample.vx + xController.calculate(pose.x, sample.x),
+      sample.vy + yController.calculate(pose.y, sample.y),
+      sample.omega + headingController.calculate(pose.rotation.radians, sample.heading)
+    )
+
+    // Apply the generated speeds
+    driveFieldRelative(speeds)
+  }
+
   fun setVoltage(volts: Double) {
     modules.forEach {
       it.setVoltage(volts)
     }
   }
 
-  fun getModuleVel(): Double {
-    var totalVel = 0.0
-    modules.forEach { totalVel += it.state.speedMetersPerSecond }
-    return totalVel / modules.size
+  /** Stops the robot's drive. */
+  fun stop() {
+    this.driveFieldRelative(ChassisSpeeds(0.0, 0.0, 0.0))
   }
 
   override fun periodic() {
@@ -88,11 +120,6 @@ open class SwerveDrive(
     )
 
     speedMagnitude = hypot(currentSpeeds.vxMetersPerSecond, currentSpeeds.vyMetersPerSecond)
-  }
-
-  /** Stops the robot's drive. */
-  fun stop() {
-    this.set(ChassisSpeeds(0.0, 0.0, 0.0))
   }
 
   /** @return An array of [SwerveModulePosition] for each module, containing distance and angle. */
